modules/getpic/pyproject.toml
[project]
name = "getpic"
version = "0.1.0"
description = "解耦式智能图表参数提取引擎，基于 MCP 协议架构"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",
    "python-multipart>=0.0.18",
    "pillow>=11.0.0",
    "opencv-python>=4.10.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"
